requirements.txt
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
flake8==7.1.1
idna==3.8
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.32.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
